+++ R/extend_family.R
@@ -1,6 +1,6 @@
 # Family-specific helper functions
 #
-# `extend_family(family)` returns a [`family`] object augmented with auxiliary
+# `extend_family(family)` returns a `family` object augmented with auxiliary
 # functions that are needed for computing KL-divergence, log predictive density,
 # dispersion projection, etc.
 #
@@ -9,13 +9,13 @@
 
 #' Extend a family
 #'
-#' This function adds some internally required elements to a [`family`] object.
-#' It is called internally by [init_refmodel()], so you will rarely need to call
-#' it yourself.
+#' This function adds some internally required elements to an object of class
+#' `family` (see, e.g., [family()]). It is called internally by
+#' [init_refmodel()], so you will rarely need to call it yourself.
 #'
-#' @param family A [`family`] object.
+#' @param family An object of class `family`.
 #'
-#' @return The [`family`] object extended in the way needed by \pkg{projpred}.
+#' @return The `family` object extended in the way needed by \pkg{projpred}.
 #'
 #' @export
 extend_family <- function(family) {