+++ R/varsel.R
@@ -1,4 +1,4 @@
-#' Variable selection (without cross-validation)
+#' Variable selection without cross-validation
 #'
 #' Perform the projection predictive variable selection for GLMs, GLMMs, GAMs,
 #' and GAMMs. This variable selection consists of a *search* part and an
@@ -77,8 +77,9 @@
 #'   results can be obtained again if needed. Passed to argument `seed` of
 #'   [set.seed()], but can also be `NA` to not call [set.seed()] at all. Here,
 #'   this seed is used for clustering the reference model's posterior draws (if
-#'   `!is.null(nclusters)`) and for drawing new group-level effects when
-#'   predicting from a multilevel submodel (however, not yet in case of a GAMM).
+#'   `!is.null(nclusters)` or `!is.null(nclusters_pred)`) and for drawing new
+#'   group-level effects when predicting from a multilevel submodel (however,
+#'   not yet in case of a GAMM).
 #' @param ... Arguments passed to [get_refmodel()] as well as to the divergence
 #'   minimizer (during a forward search and also during the evaluation part, but
 #'   the latter only if `refit_prj` is `TRUE`).
@@ -121,7 +122,7 @@
 #'   The elements of the `search_terms` character vector don't need to be
 #'   individual predictor terms. Instead, they can be building blocks consisting
 #'   of several predictor terms connected by the `+` symbol. To understand how
-#'   these building blocks works, it is important to know how \pkg{projpred}'s
+#'   these building blocks work, it is important to know how \pkg{projpred}'s
 #'   forward search works: It starts with an empty vector `chosen` which will
 #'   later contain already selected predictor terms. Then, the search iterates
 #'   over model sizes \eqn{j \in \{1, ..., J\}}{j = 1, ..., J}. The candidate
@@ -144,8 +145,8 @@
 #'   specify `search_terms = c("x2", "x3", "x2 + x3")`.
 #'
 #' @return An object of class `vsel`. The elements of this object are not meant
-#'   to be accessed directly but instead via helper functions (see the vignette
-#'   or type `?projpred`).
+#'   to be accessed directly but instead via helper functions (see the main
+#'   vignette and [projpred-package]).
 #'
 #' @seealso [cv_varsel()]
 #'