+++ tests/testthat/test_datafit.R
@@ -2,6 +2,12 @@ context("datafit")
 
 # Setup -------------------------------------------------------------------
 
+# Note: Since PR #351, offsets are not supported anymore for `datafit`s. Here,
+# we use the data as generated in `setup.R`, i.e., sometimes with offsets. So
+# far, this doesn't seem to cause problems in the submodel fitting routines, but
+# it might do in the future. Then, either the scenarios including offsets have
+# to be excluded or new data has to be generated without offsets.
+
 .extrmoddat_datafit <- function(object, newdata = NULL, wrhs = NULL,
                                 orhs = NULL, resp_form = NULL) {
   if (is.null(newdata)) {
@@ -16,16 +22,10 @@ context("datafit")
     weights <- wrhs
   }
 
-  if (inherits(orhs, "formula")) {
-    offset <- eval_rhs(orhs, newdata)
-  } else if (is.null(orhs)) {
-    offset <- newdata$offs_col
-  } else {
-    offset <- orhs
-  }
+  offset <- rep(0, nrow(newdata))
 
   if (inherits(resp_form, "formula")) {
-    y <- eval_rhs(resp_form, newdata)
+    y <- eval_el2(resp_form, newdata)
   } else {
     y <- NULL
   }
@@ -62,9 +62,6 @@ datafits <- lapply(args_datafit, function(args_datafit_i) {
     }
     if (args_datafit_i$fam_nm == "brnll") {
       newdata$wobs_col <- 1
-    }
-    if (grepl("\\.without_offs", args_datafit_i$tstsetup_fit)) {
-      newdata$offs_col <- 0
     }
     args <- nlist(object, newdata, wrhs, orhs, resp_form)
     return(do.call(.extrmoddat_datafit, args))
@@ -202,11 +199,6 @@ test_that("init_refmodel(): `object` of class \"datafit\" works", {
     } else {
       wobs_expected_crr <- rep(1, nobsv)
     }
-    if (grepl("\\.with_offs", tstsetup)) {
-      offs_expected_crr <- offs_tst
-    } else {
-      offs_expected_crr <- rep(0, nobsv)
-    }
     refmodel_tester(
       datafits[[tstsetup]],
       is_datafit = TRUE,
@@ -215,7 +207,7 @@ test_that("init_refmodel(): `object` of class \"datafit\" works", {
       formul_expected = get_formul_from_fit(fits[[tstsetup_fit]]),
       with_spclformul = with_spclformul_crr,
       wobs_expected = wobs_expected_crr,
-      offs_expected = offs_expected_crr,
+      offs_expected = rep(0, nobsv),
       nrefdraws_expected = 1L,
       fam_orig = get(paste0("f_", args_datafit[[tstsetup]]$fam_nm)),
       mod_nm = args_datafit[[tstsetup]]$mod_nm,
@@ -423,7 +415,6 @@ test_that(paste(
         tail(nobsv_tst, 1)
       ),
       weightsnew = ~ wobs_col,
-      offsetnew = ~ offs_col,
       filter_nterms = nterms_crr[1]
     )
     pl_tester(pl_with_args,
@@ -462,7 +453,6 @@ test_that(paste(
       prjs_vs_datafit[[tstsetup]],
       newdata = head(dat, tail(nobsv_tst, 1)),
       weightsnew = ~ wobs_col,
-      offsetnew = ~ offs_col,
       filter_nterms = nterms_crr[1],
       nresample_clusters = tail(nresample_clusters_tst, 1),
       .seed = seed2_tst
@@ -608,6 +598,9 @@ test_that(paste(
   "L1-projection with data reference gives the same results as",
   "Lasso from glmnet."
 ), {
+  # This test sometimes behaves inpredictably when run in `R CMD check`, so skip
+  # it on CRAN:
+  skip_on_cran()
   if (exists(".Random.seed", envir = .GlobalEnv)) {
     rng_old <- get(".Random.seed", envir = .GlobalEnv)
   }
@@ -619,7 +612,6 @@ test_that(paste(
   b <- seq(0, 1, length.out = nterms)
   dis <- runif(1, 0.3, 0.5)
   weights <- sample(1:4, n, replace = TRUE)
-  offset <- 0.1 * rnorm(n)
 
   fams <- list(gaussian(), binomial(), poisson())
   x_list <- lapply(fams, function(fam) x)
@@ -642,21 +634,6 @@ test_that(paste(
     nlist(y, y_glmnet, weights)
   })
 
-  median_lasso_preds <- list(
-    c(0.2774068, 0.2857059, 0.2878935, 0.2813947, 0.2237729,
-      0.2895152, 0.3225808, 0.3799348),
-    c(0.009607217, 0.015400719, -0.017591445, -0.009711566,
-      -0.023867036, -0.038964983, -0.036081074, -0.045065655),
-    c(1.8846845, 1.8830678, 1.8731548, 1.4232035, 0.9960167,
-      0.9452660, 0.6216253, 0.5856283)
-  )
-
-  solution_terms_lasso <- list(
-    c(10, 9, 6, 8, 7, 5, 4, 3, 1, 2),
-    c(10, 9, 8, 6, 7, 5, 3, 4, 2, 1),
-    c(9, 10, 6, 7, 3, 5, 2, 4, 3, 1)
-  )
-
   extract_model_data <- function(object, newdata = NULL, wrhs = NULL,
                                  orhs = NULL, extract_y = FALSE) {
     if (!is.null(object)) {
@@ -675,9 +652,6 @@ test_that(paste(
     if (is.null(object)) {
       if ("weights" %in% colnames(newdata)) {
         wrhs <- ~ weights
-      }
-      if ("offset" %in% colnames(newdata)) {
-        orhs <- ~ offset
       }
       if ("y" %in% colnames(newdata)) {
         resp_form <- ~ y
@@ -715,10 +689,8 @@ test_that(paste(
     ))
     expect_warning(
       pred1 <- proj_linpred(vs,
-                            newdata = data.frame(x = x, offset = offset,
-                                                 weights = weights),
+                            newdata = data.frame(x = x, weights = weights),
                             nterms = 0:nterms, transform = FALSE,
-                            offsetnew = ~offset,
                             refit_prj = FALSE),
       paste("^Currently, `refit_prj = FALSE` requires some caution, see GitHub",
             "issues #168 and #211\\.$"),
@@ -728,7 +700,6 @@ test_that(paste(
     # compute the results for the Lasso
     lasso <- glmnet::glmnet(x, y_glmnet,
                             family = fam$family, weights = weights,
-                            offset = offset,
                             lambda.min.ratio = lambda_min_ratio,
                             nlambda = nlambda, thresh = 1e-12)
     solution_terms <- predict(lasso, type = "nonzero", s = lasso$lambda)
@@ -736,32 +707,47 @@ test_that(paste(
     lambdainds <- sapply(unique(nselected), function(nterms) {
       max(which(nselected == nterms))
     })
-    ## lambdaval <- lasso$lambda[lambdainds]
-    ## pred2 <- predict(lasso,
-    ##   newx = x, type = "link", s = lambdaval,
-    ##   newoffset = offset
-    ## )
-
-    # check that the predictions agree (up to nterms-2 only, because glmnet
-    # terminates the coefficient path computation too early for some reason)
-    for (j in 1:(nterms - 2)) {
-      expect_true(median(pred1[[j]]$pred) - median_lasso_preds[[i]][j] < 3e-1)
+    lambdaval <- lasso$lambda[lambdainds]
+    pred2 <- predict(lasso, newx = x, type = "link", s = lambdaval)
+    solution_terms_lasso <- integer()
+    lasso_coefs <- as.matrix(lasso$beta[, tail(lambdainds, -1), drop = FALSE])
+    for (ii in seq_len(ncol(lasso_coefs))) {
+      solution_terms_lasso <- c(
+        solution_terms_lasso,
+        setdiff(which(lasso_coefs[, ii] > 0), solution_terms_lasso)
+      )
     }
+
+    # check that the predictions agree
+    pred1 <- unname(sapply(pred1, "[[", "pred"))
+    pred2 <- unname(pred2)
+    # Sometimes, glmnet terminates the coefficient path computation too early
+    # for some reason:
+    if (ncol(pred1) > ncol(pred2)) {
+      pred1 <- pred1[, seq_len(ncol(pred2)), drop = FALSE]
+    }
+    expect_equal(pred1, pred2, tolerance = 1e-2, info = as.character(i))
 
     # check that the coefficients are similar
     ind <- match(vs$solution_terms, setdiff(split_formula(formula), "1"))
-    if (Sys.getenv("NOT_CRAN") == "true") {
-      betas <- sapply(vs$search_path$submodls, function(x) x[[1]]$beta %||% 0)
-      delta <- sapply(seq_len(nterms), function(i) {
-        abs(t(betas[[i + 1]]) - lasso$beta[ind[1:i], lambdainds[i + 1]])
-      })
-      expect_true(median(unlist(delta)) < 6e-2)
-      expect_true(median(abs(sapply(vs$search_path$submodls, function(x) {
-        x[[1]]$alpha
-      }) - lasso$a0[lambdainds])) < 1.5e-1)
-    } else {
-      expect_true(sum(ind == solution_terms_lasso[[i]]) >= nterms / 2)
+    betas <- sapply(vs$search_path$submodls, function(x) x[[1]]$beta %||% 0)
+    delta <- sapply(seq_len(length(lambdainds) - 1), function(i) {
+      abs(t(betas[[i + 1]]) - lasso$beta[ind[1:i], lambdainds[i + 1]])
+    })
+    expect_true(median(unlist(delta)) < 6e-2)
+    expect_true(median(abs(
+      sapply(head(vs$search_path$submodls, length(lambdainds)),
+             function(x) {
+               x[[1]]$alpha
+             }) -
+        lasso$a0[lambdainds])
+    ) < 1.5e-1)
+    # Sometimes, glmnet terminates the coefficient path computation too early
+    # for some reason:
+    if (length(ind) > length(solution_terms_lasso)) {
+      ind <- ind[seq_along(solution_terms_lasso)]
     }
+    expect_identical(ind, solution_terms_lasso, info = as.character(i))
   }
   RNGversion(getRversion())
   if (exists("rng_old")) assign(".Random.seed", rng_old, envir = .GlobalEnv)